requirements.txt
# === DATABRICKS NOTEBOOKS DEPENDENCIES ===
# Dependencies for ARIA Jupyter notebooks in /src/aria/notebooks/
# Web application dependencies are in app/requirements.txt

# Core data processing (shared by notebooks)
pandas>=2.0.0
numpy>=1.21.0

# Databricks integration for notebooks
databricks-sdk==0.33.0

# Machine Learning and AI
mlflow>=2.18.0
dspy-ai>=2.5.43
databricks-agents>=0.5.0
databricks-langchain
databricks-vectorsearch

# Deep Learning and NLP
transformers>=4.44.2
torch
tokenizers

# Data Visualization and Analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotnine
numpy>=1.21.0
scipy>=1.7.0

# Document Processing
python-docx>=0.8.11
openpyxl>=3.0.0
diff-match-patch>=20200713

# Web Scraping and Browser Automation
playwright>=1.40.0
lxml>=4.9.0

# Language Model Integration
langchain>=0.1.13
openai

# Note: Some packages may already be pre-installed in Databricks environments
# but are listed here for completeness and local development compatibility
#
# For web application dependencies, see: app/requirements.txt